attached_assets/requirements_1752618908971.txt
fastapi
uvicorn[standard]
python-dotenv
twilio
websockets
httpx
openai
jq
